updated gulpfile and fixed styleguide generator
