+++ gulpfile.babel.js
@@ -4,7 +4,6 @@ import plugins from 'gulp-load-plugins';
 import yargs from 'yargs';
 import browser from 'browser-sync';
 import gulp from 'gulp';
-import panini from 'panini';
 import del from 'del';
 import sherpa from 'style-sherpa';
 import yaml from 'js-yaml';
@@ -26,7 +25,7 @@ function loadConfig () {
 
 // Build the "dist" folder by running all of the below tasks
 gulp.task('build',
- gulp.series(clean, gulp.parallel(pages, sass, javascript, images, copy), styleGuide));
+gulp.series(clean, gulp.parallel(metalsmith, sass, javascript, images, copy), styleGuide));
 
 // Build the site, run the server, and watch for file changes
 gulp.task('default',
@@ -45,30 +44,92 @@ function copy () {
     .pipe(gulp.dest(PATHS.dist + '/assets'));
 }
 
-// Copy page templates into finished HTML files
-function pages () {
-  return gulp.src('src/pages/**/*.{html,hbs,handlebars}')
-    .pipe(panini({
-      root: 'src/pages/',
-      layouts: 'src/layouts/',
-      partials: 'src/partials/',
-      data: 'src/data/',
-      helpers: 'src/helpers/'
-    }))
-    .pipe(gulp.dest(PATHS.dist));
+// reads the content of package.json
+function getPackageJson () {
+  return JSON.parse(fs.readFileSync('./package.json', 'utf8'));
 }
 
-// Load updated HTML templates and partials into Panini
-function resetPages (done) {
-  panini.refresh();
-  done();
+/** run metalsmith (static site generator)
+ * if `PRODUCTION` is set to true, additionally append the package.json version
+ * to the css file
+ */
+function metalsmith () { // eslint-disable-line no-unused-vars
+  // Metalsmith-specific requires
+  const gulpsmith = require('gulpsmith');
+  const frontMatter = require('gulp-front-matter');
+  const assign = require('lodash.assign');
+  const handlebars = require('handlebars');
+  const htmlMinifier = require('metalsmith-html-minifier');
+  const markdown = require('metalsmith-markdown');
+  const layouts = require('metalsmith-layouts');
+  const pageTitles = require('metalsmith-page-titles');
+  const permalinks = require('metalsmith-permalinks');
+  const sitemap = require('metalsmith-mapsite');
+  const debug = require('metalsmith-debug');
+
+  // filter out files with front matter
+  var fmFilter = $.filter('**/*.{html,md,htb}', { restore: true });
+
+  // reget package
+  var pkg = getPackageJson();
+
+  // register Handlebars helpers
+  handlebars.registerHelper('moment', require('helper-moment'));
+  require('swag').registerHelpers(handlebars);
+  // custom helpers
+  handlebars.registerHelper('ext2Png', function (str) {
+    return str.substr(0, str.lastIndexOf('.')) + '.png';
+  });
+  handlebars.registerHelper('ext2Jpg', function (str) {
+    return str.substr(0, str.lastIndexOf('.')) + '.jpg';
+  });
+
+  return gulp.src('./src/pages/**/*')
+    .pipe(fmFilter)
+     // grab files with front matter and assign them as a property so metalsmith will find it
+    .pipe(frontMatter({
+      property: 'frontMatter'
+    })).on('data', function (file) {
+      assign(file, file.frontMatter);
+      delete file.frontMatter;
+    })
+    // remove the filter (back to everything in /src) and let metalsmith do its thing
+    .pipe(fmFilter.restore)
+    .pipe(
+      gulpsmith()
+        .metadata({
+          'site': {
+            'title': CONST.title
+          }
+        })
+        .use(pageTitles())
+        .use(markdown())
+        .use(permalinks(':collection/:title'))
+        .use(layouts({
+          'engine': 'handlebars',
+          'directory': PATHS.templates,
+          'partials': PATHS.templates + '/partials'
+        }))
+        .use(sitemap({
+          'hostname': 'http://stateofthebrowser.com',
+          'changefreq': 'yearly',
+          'pattern': [ '{illustration,photo,pixelart}/**', 'about/*' ], // FIXME do something here
+          'omitIndex': true
+        }))
+        .use(htmlMinifier())
+        .use(debug())
+   )
+    .pipe($.if(PRODUCTION, $.replace('.css', '.' + pkg.version + '.css')))
+    // .pipe($.if(PRODUCTION, replace(/(main|vendor|modernizr)\.js/g, '$1.' + pkg.version + '.js')))
+    .pipe(gulp.dest('./' + PATHS.dist))
+    .pipe(browser.reload({ stream: true }));
 }
 
 // Generate a style guide from the Markdown content and HTML template in styleguide/
 function styleGuide (done) {
   sherpa('src/styleguide/index.md', {
     output: PATHS.dist + '/styleguide.html',
-    template: 'src/styleguide/template.html'
+    template: 'src/templates/styleguide.hbs'
   }, done);
 }
 
@@ -133,8 +194,8 @@ function reload (done) {
 // Watch for changes to static assets, pages, Sass, and JavaScript
 function watch () {
   gulp.watch(PATHS.assets, copy);
-  gulp.watch('src/pages/**/*.html').on('all', gulp.series(pages, browser.reload));
-  gulp.watch('src/{layouts,partials}/**/*.html').on('all', gulp.series(resetPages, pages, browser.reload));
+  gulp.watch('src/pages/**/*.html').on('all', gulp.series(metalsmith, browser.reload));
+  // gulp.watch('src/{layouts,partials}/**/*.html').on('all', gulp.series(resetPages, pages, browser.reload));
   gulp.watch('src/assets/scss/**/*.scss').on('all', sass);
   gulp.watch('src/assets/js/**/*.js').on('all', gulp.series(javascript, browser.reload));
   gulp.watch('src/assets/img/**/*').on('all', gulp.series(images, browser.reload));
